[minor] refactoring + parallel processing unit test fixed
